requirements.txt
flask
requests
python-aqi
